reverted changes
Realized that trying to make predictions from 0-10 with one vs all
logistic regression was near impossible because logistic regression is
good for fitting binary outcomes — T/F. One vs all was just comparing 1
number out of the rest of the numbers, and wasn’t able to give a
specific prediction as to what number it was looking at. Onwards with
neural networks.